+++ lib/api/v3/work_packages/schema/work_package_schema_representer.rb
@@ -33,7 +33,7 @@ require 'roar/json/hal'
 module API
   module V3
     module WorkPackages
-      module Form
+      module Schema
         class WorkPackageSchemaRepresenter < ::API::Decorators::Single
 
           def self.property_schema(property,
